wire up okhttp bug handler
